Create a new grunt task `grunt pack`
It generates the npm modules as we would like to use in other apps.

The idea is to compile source file in ES6 or coffee to pure javascript (ES5).
So we take files from `src/react-components` and compile them to `dist/lib`
Then we generate the `dist/mnd.js` file which “require” all components.

So that you can use the components like this in another app:

```
DatePicker = require(‘mnd’).DatePicker
Dropdown = require(‘mnd’).Dropdown
```

Not that the component name (i.e.: DatePicker) comes from the file name, converted to PascalCase, in this case: date_picker => DatePicker